fix(csharp/src/Drivers/Databricks): Correct DatabricksCompositeReader and StatusPoller to Stop/Dispose Appropriately (#3217)
### Motivation
The following cases are not properly stopping or disposing the status
poller:
1.  If the DatabricksCompositeReader is explicitly disposed by the user
2. CloudFetchReader is done returning results
3. Edge case terminal operation status (timedout_state, unknown_state)

In addition:
- When DatabricksOperationStatusPoller.Dispose(), it may cancel the
GetOperationStatusRequest in the client. If the input buffer has data
and cancellation is triggered, it leaves the TCLI client with
unconsumed/unsent data in the buffer, breaking subsequent requests
(fixed in this PR)

### Fixes

DatabricksOperationStatusPollerLogic is now more appropriately managed
by DatabricksCompositeReader (moved out of BaseDatabricksReader) to
handle all cases where null results (indicating completion) are
returned.

Disposing DatabricksCompositeReader appropriately disposes the
activeReader and statusPoller


#### TODO
Follow-up PR - when statement is disposed, it should also dispose the
reader (the poller is currently stopped when operationhandle is set to
null, but this should also happen explicitly)

Need add some unit testing (follow up pr:
https://github.com/apache/arrow-adbc/pull/3243)