feat(csharp/src/Drivers): instrument tracing exporters for BigQuery/Apache drivers (#3315)
Adds instrumentation to support tracing exporters

# Traces Listeners

## FileActivityListener

Provides an ActivityListener to write telemetry traces to
rotating files in folder. File names are created with the following
pattern:
`<trace-source>-<YYYY-MM-DD-HH-mm-ss-fff>-<process-id>.log`.

For example:
`apache.arrow.adbc.drivers.databricks-2025-08-15-10-35-56-012345-99999.log`.

The default folder used is:

| Platform | Folder |
| --- | --- |
| Windows | `%LOCALAPPDATA%/Apache.Arrow.Adbc/Traces` |
| macOS | `$HOME/Library/Application Support/Apache.Arrow.Adbc/Traces` |
| Linux   | `$HOME/.local/share/Apache.Arrow.Adbc/Traces` |

By default, up to 999 files of maximum size 1024 KB are written to
the trace folder.

The environment variable `OTEL_TRACES_EXPORTER` can be used to select
one of the
available exporters. Or the database parameter `adbc.traces.exporter`
can be used,
which has precedence over the environment variable.

The following listeners are supported:

| Listener | Description |
| --- | --- |
| `adbcfile` | Exports traces to rotating files in a folder. |

The `FileActivityListener` is designed to allow an instance to be
associated with and have the same lifespan as a `Connection` instance.